Streamlined build to always exclude external dependencies.
