+++ .github/workflows/npm.yaml
@@ -29,7 +29,7 @@ jobs:
         with:
           script: core.setFailed('Release tag does not match package.json')
 
-      - run: yarn build:nodeps
+      - run: yarn build
 
       - name: Publish with latest tag
         if: github.event.release.prelease == false
